moved clinvar submission to the institute level
