+++ scout/server/blueprints/cases/controllers.py
@@ -7,7 +7,7 @@ import os
 import query_phenomizer
 import requests
 from bs4 import BeautifulSoup
-from flask import current_app, url_for, flash, redirect
+from flask import current_app, url_for, flash
 from flask_login import current_user
 from flask_mail import Message
 from xlsxwriter import Workbook
@@ -29,7 +29,6 @@ from scout.constants.variant_tags import (
     MANUAL_RANK_OPTIONS,
 )
 from scout.export.variant import export_mt_variants
-from scout.parse.clinvar import clinvar_submission_header, clinvar_submission_lines
 from scout.parse.matchmaker import (
     genomic_features,
     hpo_terms,
@@ -411,115 +410,6 @@ def coverage_report_contents(store, institute_obj, case_obj, base_url):
     coverage_data = "".join(["%s" % x for x in soup.body.contents])
 
     return coverage_data
-
-
-def update_clinvar_submission_status(store, request, institute_id, submission_id):
-    """Update the status of a clinVar submission
-
-    Args:
-        store(adapter.MongoAdapter)
-        request(flask.request) POST request sent by form submission
-        institute_id(str) institute id
-        submission_id(str) the database id of a clinvar submission
-    """
-    update_status = request.form.get("update_submission")
-
-    if update_status == "close":  # close a submission
-        store.update_clinvar_submission_status(institute_id, submission_id, "closed")
-    elif update_status == "open":
-        store.update_clinvar_submission_status(
-            institute_id, submission_id, "open"
-        )  # open a submission
-    elif update_status == "register_id" and request.form.get(
-        "clinvar_id"
-    ):  # provide an official clinvar submission ID
-        result = store.update_clinvar_id(
-            clinvar_id=request.form.get("clinvar_id"), submission_id=submission_id,
-        )
-    elif request.form.get("update_submission") == "delete":  # delete a submission
-        deleted_objects, deleted_submissions = store.delete_submission(submission_id=submission_id)
-        flash(
-            f"Removed {deleted_objects} objects and {deleted_submissions} submission from database",
-            "info"
-        )
-
-
-def clinvar_submissions(store, institute_id):
-    """Get all Clinvar submissions for a user and an institute"""
-    submissions = list(store.clinvar_submissions(institute_id))
-    return submissions
-
-
-def _generate_csv(header, lines):
-    yield header + "\n"
-    for line in lines:  # lines have already quoted fields
-        yield line + "\n"
-
-
-def clinvar_submission_file(store, request, submission_id):
-    """Prepare content (header and lines) of a csv clinvar submission file
-
-    Args:
-        store(adapter.MongoAdapter)
-        request(flask.request) POST request sent by form submission
-        submission_id(str) the database id of a clinvar submission
-
-    Returns:
-        (filename, csv_header, csv_lines):
-            filename(str) name of file to be downloaded
-            csv_header(list) string list content of file header
-            csv_lines(list) string list content of file lines
-    """
-    clinvar_subm_id = request.form.get("clinvar_id")
-    if clinvar_subm_id == "":
-        flash(
-            "In order to download a submission CSV file you should register a Clinvar submission Name first!",
-            "warning",
-        )
-        return
-
-    csv_type = request.form.get("csv_type", "")
-
-    submission_objs = store.clinvar_objs(
-        submission_id=submission_id, key_id=csv_type
-    )  # a list of clinvar submission objects (variants or casedata)
-
-    if submission_objs is None or len(submission_objs) == 0:
-        flash(
-            f"There are no submission objects of type '{csv_type}' to include in the csv file!",
-            "warning")
-        return
-
-    # Download file
-    csv_header_obj = clinvar_header(
-        submission_objs, csv_type
-    )  # custom csv header (dict as in constants CLINVAR_HEADER and CASEDATA_HEADER, but with required fields only)
-    csv_lines = clinvar_lines(
-        submission_objs, csv_header_obj
-    )  # csv lines (one for each variant/casedata to be submitted)
-    csv_header = list(csv_header_obj.values())
-    csv_header = [
-        '"' + str(x) + '"' for x in csv_header
-    ]  # quote columns in header for csv rendering
-
-    today = str(datetime.datetime.now().strftime("%Y-%m-%d"))
-    filename = f"{clinvar_subm_id}_{csv_type}_{today}.csv"
-
-    return (filename, csv_header, csv_lines)
-
-
-def clinvar_header(submission_objs, csv_type):
-    """ Call clinvar parser to extract required fields to include in csv header from clinvar submission objects"""
-
-    clinvar_header_obj = clinvar_submission_header(submission_objs, csv_type)
-    return clinvar_header_obj
-
-
-def clinvar_lines(clinvar_objects, clinvar_header):
-    """ Call clinvar parser to extract required lines to include in csv file from clinvar submission objects and header"""
-
-    clinvar_lines = clinvar_submission_lines(clinvar_objects, clinvar_header)
-    return clinvar_lines
 
 
 def mt_excel_files(store, case_obj, temp_excel_dir):
