+++ scout/server/blueprints/cases/templates/cases/causatives.html
@@ -29,7 +29,7 @@
     SNVs and INDELs</a>
 </li>
 <li class="nav-item">
-  <a class="nav-link" href="{{ url_for('cases.clinvar_submissions', institute_id=institute._id) }}">Clinvar
+  <a class="nav-link" href="{{ url_for('overview.clinvar_submissions', institute_id=institute._id) }}">Clinvar
     submissions</a>
 </li>
 {% endblock %}
