+++ scout/server/blueprints/cases/templates/cases/gene_variants.html
@@ -25,7 +25,7 @@
      <span class="navbar-text">All SNVs and INDELs</span>
    </li>
    <li class="nav-item">
-     <a class="nav-link" href="{{ url_for('cases.clinvar_submissions', institute_id=institute._id) }}">Clinvar submissions</a>
+     <a class="nav-link" href="{{ url_for('overview.clinvar_submissions', institute_id=institute._id) }}">Clinvar submissions</a>
    </li>
 
 {% endblock %}
