+++ scout/server/blueprints/institutes/templates/overview/institute.html
@@ -21,6 +21,9 @@
   <li class="nav-item active">
     <p class="navbar-text text-nowrap">{{ institute.display_name }}</p>
   </li>
+  <li class="nav-item">
+    <a class="nav-link" href="{{ url_for('overview.clinvar_submissions', institute_id=institute._id) }}">Clinvar submissions</a>
+  </li>
 {% endblock %}
 
 {% block content_main %}
