+++ scout/server/blueprints/variant/templates/variant/variant.html
@@ -333,7 +333,7 @@
             class="btn btn-outline-secondary form-control" target="_blank">Submit to ClinVar</a>
         {% endif %}
         {% if variant.clinvar_clinsig %}
-          <a href="{{ url_for('cases.clinvar_submissions', institute_id=institute._id) }}" class="btn btn-outline-secondary form-control">Modify clinvar submission</a>
+          <a href="{{ url_for('overview.clinvar_submissions', institute_id=institute._id) }}" class="btn btn-outline-secondary form-control">Modify clinvar submission</a>
         {% endif %}
       </div>
       {% if evaluations %}
